Remove some signed/unsigned integer comparison warnings (#3643)
* Remove some signed/unsigned integer comparison warnings

Update executor.cc

Leaving this modification apart, as requested by @rmlarsen #3643

* Update array_ops.cc

Patching a messy rebase

+++ tensorflow/core/framework/tensor.cc
@@ -740,7 +740,7 @@ string Tensor::SummarizeValue(int64 max_entries) const {
       string ret;
       // TODO(irving): Don't call flat every time around this
       // loop.
-      for (int64 i = 0; i < limit; ++i) {
+      for (size_t i = 0; i < limit; ++i) {
         if (i > 0) strings::StrAppend(&ret, " ");
         switch (dtype()) {
           case DT_STRING:
